messages: AsciiImage + AsciiImageSection + EmojiCardBg + DanaLabel (batch 1)
Port canvas/ascii components from React → Svelte 5.
- src/lib/deana/images.ts: moved from components/messages/deana-images.ts
- DanaLabel.svelte: $state + setInterval (onMount)
- AsciiImage.svelte: thin wrapper delegating to asciiImage action
- actions/asciiImage.ts: full canvas/crossfade logic as Svelte action
- AsciiImageSection.svelte: LazyMount + AsciiImage + DanaLabel layout
- EmojiCardBg.svelte: thin wrapper around AsciiGallery

Deleted: components/messages/{ascii-image,ascii-image-section,emoji-card-bg,dana-label}.tsx + deana-images.ts
