+++ src/lib/deana/images.ts
@@ -1,0 +1,8 @@
+export const DEANA_IMAGES = [
+	"/assets/deana-6.jpg",
+	"/assets/deana-5.webp",
+	"/assets/deana-hero-3.webp",
+	"/assets/deana-hero.webp",
+	"/assets/deana-hero-5.webp",
+	"/assets/deana-hero-6.webp",
+];
